feat: add getError util and extra constants - implement to accept bid and set buy now
