+++ src/constants/errors.ts
@@ -1,0 +1,15 @@
+export const ERRORS = {
+	REJECTED_WALLET: 'Rejected by wallet.',
+	DATA_CONSUMED:
+		'The bid you are trying to accept is no longer valid. Please select another bid to accept',
+	PROBLEM_OCCURRED: 'An unexpected problem has occurred. Please try again.',
+	SIGNATURE: 'Failed to generate signature.',
+	TRANSACTION: 'Transaction failed. Please try again',
+	LIBRARY: 'Failed to connect with Web3 wallet.',
+	CONSOLE_TEXT: 'Failed to check zAuction approval status',
+};
+
+export const MESSAGES = {
+	TRANSACTION_DENIED: 'denied transaction signature',
+	DATA_CONSUMED: 'data already consumed',
+};
