+++ src/containers/flows/AcceptBid/AcceptBid.constants.tsx
@@ -43,21 +43,9 @@ export const STATUS_TEXT = {
 	APPROVED: 'Approved',
 };
 
-export const ERRORS = {
-	SIGNATURE: 'Failed to generate signature.',
-	TRANSACTION: 'Transaction failed. Please try again',
-	LIBRARY: 'Failed to connect with Web3 wallet.',
-	CONSOLE_TEXT: 'Failed to check zAuction approval status',
-	REJECTED_WALLET: 'Rejected by wallet.',
-	DATA_CONSUMED:
-		'The bid you are trying to accept is no longer valid. Please select another bid to accept',
-	PROBLEM_OCCURRED: 'An unexpected problem has occurred. Please try again.',
-};
-
 export const MESSAGES = {
 	TEXT_LOADING: 'Loading Bid Data...',
 	TEXT_FAILED_TO_LOAD: 'Failed to load bid data.',
-	TRANSACTION_DENIED: 'denied transaction signature',
 	TEXT_CONFIRMATION:
 		'This transaction is about to be seared upon the blockchain. There’s no going back.',
 	TEXT_ACCEPT_PROMPT: 'Please accept wallet transaction...',
@@ -68,8 +56,6 @@ export const MESSAGES = {
 		'Waiting for transaction approval. You should receive a request in your wallet.',
 	SUCCESS_CONFIRMATION: 'Success! Bid accepted and ownership transferred',
 	CONFIRM_BID_AMOUNT: 'Are you sure you want to accept a bid of',
-	DATA_CONSUMED: 'data already consumed',
-	INCORRECT_BIDDER: 'recovered incorrect bidder',
 };
 
 export const getConfirmNFTPriceDetails = (
